IMAGE: Loading libskiahw or ARM codec depending of a dynamic variable.
Three check conditions have been added in this code:
1. OMAP4 Macro
2. jpeg.libskiahw.decoder.enable
3. jpeg.libskiahw.decoder.thresh

It's been used OMAP4 macro to ensure this change is applicable only to OMAP4
in other case it will keep doing the same.

Depending of jpeg.libskiahw.decoder variable it's been choose
between ARM and libskiahw codec to be loaded.

The threshold variable jpeg.libskiahw.decoder.thresh has been defined for future
purposes so we can choose the codec depending of the size of the image, but right
now is set up as zero, making no effect in the flow of the code.

TARGET_OMAP4 macro was added on Android.mk file.

Change-Id: I9d18f167487a11df4f4557e5760b17dcfa5f21d9
Signed-off-by: Octaviano Flores <x0083028@ti.com>