OMAP4:SKIA: Handle decoding of premature jpeg files.
[Comment]
In SW jpeg decoder, the premature jpeg file decoding is not handled
properly. So the images were not displayed using the android gallery3d
application.
This patch will handle the decoding of premature jpeg files by sending
dummy EOI marker to the decoder.
This patch fixes the DR OMAPS00217642.

[DR]
OMAPS00217642

Change-Id: I146c8cc260e4fddb7c1d59dfe4d941c226bdba27
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>