OMAP4:SKIA: Handle decoding of premature jpeg files.
[Comment]
In SW jpeg decoder, the premature jpeg file decoding is not handled
properly. So the images were not displayed using the android gallery3d
application.
This patch will handle the decoding of premature jpeg files by sending
dummy EOI marker to the decoder.

This patch is porting of patch-7897 from p-froyo branch.

[DR]
-

Change-Id: I1139ab9f41d3c8f82521ef7add169a81df4680d1
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>