skia: Handles decoding of premature jpeg files.
In SW jpeg decoder, the premature jpeg file decoding is not handled
properly. So the images were not displayed using the android gallery3d
application.
This patch will handle the decoding of premature jpeg files by sending
dummy EOI marker to the decoder.

This patch was ported from gb to ics and now is porting to jb.
Source: http://review.omapzoom.org/10250

Change-Id: Ic8d5f0d32c5eb06293d73bc7074526268368c1a2
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>
Signed-off-by: Andrii Beregovenko <a.beregovenko@ti.com>