SKIA: Handles decoding of premature jpeg files.
[Comment]
In SW jpeg decoder, the premature jpeg file decoding is not handled
properly. So the images were not displayed using the android gallery3d
application.
This patch will handle the decoding of premature jpeg files by sending
dummy EOI marker to the decoder.

This patch is porting from p-gingerbread to ics.
http://review.omapzoom.org/#change,10250

[DR]
-

Change-Id: Ic8d5f0d32c5eb06293d73bc7074526268368c1a2
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>